vendor: Switch the search bar round on click
We should thank google every day of our life for doing stuff in a horrible way. This may be a really ugly method but this saves us from a unneccesary fork.
Let's use the same value of the search bar overlay to keep it consistent. And also modify the search_bar_margin to use the same values as Settings

Change-Id: If76f021d418171a9126ddc10a75a63bd8d967a04
Signed-off-by: Srijith Bharadwaj <srijithdamerla2001@gmail.com>